Add files via upload
CUSTOMER PERSONALITY ANALYSIS

• Customer Personality Analysis is a detailed analysis of a company’s ideal
customers. It helps a business to better understand its customers and makes it
easier for them to modify products according to the specific needs, behaviors and
concerns of different types of customers.
• Customer personality analysis helps a business to modify its product based on its
target customers from different types of customer segments. For example, instead
of spending money to market a new product to every customer in the company’s
database, a company can analyze which customer segment is most likely to buy
the product and then market the product only on that particular segment.